requirements.txt
aniso8601==9.0.1
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudinary==1.25.0
colorama==0.4.4
cors==1.0.1
dj-database-url==0.5.0
Django==3.2.3
django-extensions==3.1.3
Django-Verify-Email==1.0.7
dnspython==2.1.0
filelock==3.0.12
Flask==2.0.1
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.9
future==0.18.2
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Pillow==8.2.0
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
PySocks==1.7.1
python-decouple==3.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.1
tldextract==3.1.0
urllib3==1.26.5
Werkzeug==2.0.1
whitenoise==5.2.0
zope.event==4.5.0
zope.interface==5.4.0
